requirements.txt
anyio==3.6.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
fastapi==0.89.1
firebase-admin==6.1.0
fonttools==4.38.0
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.14.0
httplib2==0.21.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.3
msgpack==1.0.5
numpy==1.24.2
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
packaging==23.0
pathlib==1.0.1
Pillow==9.4.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-multipart==0.0.5
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.20.0